Updated form urls + better submission handling
Views.py:

I made the file redirect to the new updated url for the thank-you page and thats pretty much it

urls.py:

I added a name for the thank-you page (thank_you) so that its easier to access this site in the views.py file

forms.py:

I added a new field with a drop down menu (ticket_num) which allows the user to select the number of tickets they want to buy at a time (if they want to buy for friends or something). I also added additional form validation for the fields so that users can't type invalid stuff into the form. The submission button has a different title now that shows the price of the tickets as $5 (subject to change based on what we decide later).