Updated form URL
I made it so that there is a /payments/ after whatever the website name is (i.e. stadium/payments/ instead of just stadium/ for the form url).
This change also means that the url for the thank-you page is now payments/thank-you rather than just /thank-you (i.e. stadium/payments/thank-you/ rather than just stadium/thank-you)

We can change the name of the thank-you url to something else later since we only have to edit it in 1 place (payments app urls.py file)